requirements.txt
# ==============================================================
# Multimodal Psychophysiological Analysis System
# Python Dependencies
# ==============================================================

# ── Phase 1 (rPPG — classical signal processing) ──────────────
opencv-python>=4.8.0         # Video capture, image warping, drawing
mediapipe>=0.10.14            # Face Mesh landmark detection
numpy>=1.24.0                 # Numerical arrays and FFT
scipy>=1.11.0                 # Butter filter, filtfilt, find_peaks
matplotlib>=3.7.0             # Real-time plotting (interactive mode)
PyYAML>=6.0                   # config.yaml loading

# ── Phase 2 (emotion recognition + fusion) ────────────────────
torch>=2.1.0                  # Deep learning framework
torchvision>=0.16.0           # Pre-trained vision models (ResNet, EfficientNet)
pandas>=2.0.0                 # DataFrame for annotations and results
openpyxl>=3.1.0               # Excel file reading (SAMM annotations)
scikit-learn>=1.3.0           # Metrics (F1, accuracy, confusion matrix)
tqdm>=4.65.0                  # Progress bars for training

# ── Optional (advanced features) ──────────────────────────────
# albumentations>=1.3.1       # Advanced data augmentation
# tensorboard>=2.14.0         # Training visualization
# timm>=0.9.0                  # Additional pre-trained models
